Core_Components/news_monitoring_agent/webserver/pyproject.toml
[project]
name = "webserver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.120.2",
    "feedparser>=6.0.12",
    "googlenews>=1.6.15",
    "lxml-html-clean>=0.4.3",
    "newspaper4k>=0.9.3.1",
    "torch>=2.9.0",
    "transformers>=4.57.1",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]
